Address review: store-owned reconcile state, drop test-only debug seams
- Replace the Objective-C associated-object state (and its top-level
  association key) with a declared DockSplitStore-owned
  dockPortalReconcileState property, per the no-ambient-global-state
  review rule. DockSplitStore.swift stays under its length budget by
  relocating the focus-ownership browserPanel(owning:in:) lookup to
  DockSplitStore+PaneFocus.swift.
- Remove the #if DEBUG debug…ForTesting extension, per the
  no-test-debug-seam-in-production-source rule. Tests now drive the
  internal reconcileDockPortalPass(reason:)/clearDockPortalReconcile()
  API and read the store-owned state directly via @testable import.
