+++ Sources/DockSplitStore+PaneFocus.swift
@@ -6,6 +6,24 @@ extension DockSplitStore {
         AppDelegate.shared?.noteRightSidebarKeyboardFocusIntent(mode: .dock, in: window)
     }
 
+    func browserPanel(owning responder: NSResponder?, in window: NSWindow?) -> BrowserPanel? {
+        guard let responder, let window else { return nil }
+        if let focused = focusedPanelId,
+           let browser = panels[focused] as? BrowserPanel,
+           browser.ownedFocusIntent(for: responder, in: window) != nil {
+            return browser
+        }
+        for (panelId, panel) in panels {
+            guard panelId != focusedPanelId,
+                  let browser = panel as? BrowserPanel,
+                  browser.ownedFocusIntent(for: responder, in: window) != nil else {
+                continue
+            }
+            return browser
+        }
+        return nil
+    }
+
     func focusedDockPaneSelection() -> (pane: PaneID?, tab: TabID?) {
         let pane = bonsplitController.focusedPaneId
         return (pane, pane.flatMap { bonsplitController.selectedTab(inPane: $0)?.id })
